Update eco benefits to be more modular
In particular, benefits are now provided by subclasses of
BenefitCalculator for both bulk and single item benefits.

The main gist is that each BenefitCalculator returns benefits scoped by
group. For instance, the tree benefit calculator may return:

{ 'plants':
   { 'stormwater': { 'value': 5, ... },
     'co2storage': { 'value': 9, ... }}}

The benefits system will then automatically merge keys that are the same
across groups. For example, if we add the ability to plant shrubs and
their eco benefits return:

{ 'plants':
   { 'stormwater': { 'value': 2, ... }}}

The summed benefit list (with the example tree and shrub) would be:

{ 'plants':
   { 'stormwater': { 'value': 7, ... },
     'co2storage': { 'value': 9, ... }}}